refactor(issues): extract descriptionPreview into its own module
Board cards were the only surface flattening description Markdown, so the
helper lived in board-card.tsx and was exported for its test. Give it a home
of its own so any surface showing a description snippet can share it.
